Refactor install into arch specific files, fix byte array length issue for darwin
